requirements-e2e.txt
# E2E test dependencies (cross-cutting, not inside any package)
playwright>=1.40
pytest-timeout>=2.2
